Extract `SettingsUpgrader` to its own file
