+++ legacy/core/src/main/java/com/fsck/k9/preferences/IdentitySettingsDescriptions.java
@@ -13,7 +13,6 @@ import com.fsck.k9.EmailAddressValidator;
 import com.fsck.k9.preferences.Settings.BooleanSetting;
 import com.fsck.k9.preferences.Settings.InvalidSettingValueException;
 import com.fsck.k9.preferences.Settings.SettingsDescription;
-import com.fsck.k9.preferences.Settings.SettingsUpgrader;
 import com.fsck.k9.preferences.Settings.V;
