+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/AccountSettingsUpgraderTo80.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.notification.NotificationLight;
 import com.fsck.k9.notification.NotificationLightDecoder;
-import com.fsck.k9.preferences.Settings.SettingsUpgrader;
+import com.fsck.k9.preferences.SettingsUpgrader;
 
 
 /**
