+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/GeneralSettingsUpgraderTo24.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import app.k9mail.legacy.preferences.AppTheme;
 import app.k9mail.legacy.preferences.SubTheme;
-import com.fsck.k9.preferences.Settings.SettingsUpgrader;
+import com.fsck.k9.preferences.SettingsUpgrader;
 
 
 /**
