+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/GeneralSettingsUpgraderTo58.java
@@ -4,7 +4,7 @@ package com.fsck.k9.preferences.upgrader;
 import java.util.Map;
 
 import app.k9mail.legacy.preferences.AppTheme;
-import com.fsck.k9.preferences.Settings.SettingsUpgrader;
+import com.fsck.k9.preferences.SettingsUpgrader;
 
 
 /**
