+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/GeneralSettingsUpgraderTo89.kt
@@ -1,12 +1,12 @@
 package com.fsck.k9.preferences.upgrader
 
-import com.fsck.k9.preferences.Settings.SettingsUpgrader
+import com.fsck.k9.preferences.SettingsUpgrader
 
 /**
  * Combine `messageViewReturnToList` and `messageViewShowNext` into `messageViewPostDeleteAction`.
  */
 class GeneralSettingsUpgraderTo89 : SettingsUpgrader {
-    override fun upgrade(settings: MutableMap<String, Any>) {
+    override fun upgrade(settings: MutableMap<String, Any?>) {
         val messageViewReturnToList = settings["messageViewReturnToList"] as? Boolean
         val messageViewShowNext = settings["messageViewShowNext"] as? Boolean
