+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/ServerSettingsUpgraderTo92.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.preferences.upgrader
 
 import com.fsck.k9.preferences.ServerSettingsDescriptions.Companion.CONNECTION_SECURITY
-import com.fsck.k9.preferences.Settings.SettingsUpgrader
+import com.fsck.k9.preferences.SettingsUpgrader
 
 class ServerSettingsUpgraderTo92 : SettingsUpgrader {
     override fun upgrade(settings: MutableMap<String, Any?>) {
