+++ legacy/core/src/main/java/com/fsck/k9/preferences/upgrader/ServerSettingsUpgraderTo94.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.preferences.upgrader
 
 import com.fsck.k9.preferences.ServerSettingsDescriptions.Companion.AUTHENTICATION_TYPE
 import com.fsck.k9.preferences.ServerSettingsDescriptions.Companion.CONNECTION_SECURITY
-import com.fsck.k9.preferences.Settings.SettingsUpgrader
+import com.fsck.k9.preferences.SettingsUpgrader
 
 /**
  * Removes legacy authentication values.
