+++ legacy/core/src/test/java/com/fsck/k9/preferences/SettingsTest.kt
@@ -8,7 +8,6 @@ import assertk.assertions.isFalse
 import assertk.assertions.isInstanceOf
 import com.fsck.k9.preferences.Settings.BooleanSetting
 import com.fsck.k9.preferences.Settings.SettingsDescription
-import com.fsck.k9.preferences.Settings.SettingsUpgrader
 import com.fsck.k9.preferences.Settings.StringSetting
 import java.util.TreeMap
 import kotlin.test.Test
